more work on weapons
added a magazine size to the weapons script, and some file management
