+++ Assets/Scripts/UI Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: aaab64608759cb844b2adab6f1db8040
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
