+++ Assets/Scripts/UI Scripts/Crosshair.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b4e30a66f1e55f64c820bdb7bfbef7b1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
