+++ Assets/Scripts/UI Scripts/HealthBarUi.cs
@@ -1,0 +1,13 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class HealthBarUi : MonoBehaviour
+{
+    [SerializeField]
+    private UnityEngine.UI.Image hpForeGround;
+    public void UpdateHealthBar(PlayerHealthController hc)
+    {
+        hpForeGround.fillAmount = hc.healthPercent;
+    }
+}
