+++ Assets/Scripts/UI Scripts/HealthBarUi.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9d0b74cbcb8ada44e919a607147c3f13
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
